requirements.txt
flet
pycryptodome
